more req on normalization and aggregation functions
Address this comments: 

5.3

   In the case of service contact instances and C-SMAs are provided by
   different vendors, it is needed to use the same common normalization
   function and aggregation functions, so that the service contact
   instance selection result can be fair among all the service contact
   instances.

This is very true, but this is not all! You need the same normalization across all components. I.e., all service contact instances and all C-SMAs in the CATS network need to do the same normalization otherwise the reported metrics will be inconsistent. And that leaves a piece you must describe which is that the normalization method must be one of:
- standardised
- configurable
- distributed