(5.2) clarification about centralized control "plane"
5.2

   *  *Distributed model*:  Computing metrics are distributed among
         network devices directly using distributed protocols without
         interactions with a centralized control plane.  Service
         scheduling function is performed by the CATS-Forwarders in the
         distribution model, therefore, the C-PS is integrated into an
         Ingress CATS-Forwarder.

   *  *Centralized model*:  Computing metrics are collected by a
         centralized control plane, and then the centralized control
         plane computes the forwarding path for service requests and
         syncs up with the Ingress CATS-Forwarder.  In this model, C-PS
         is implemented in the centralized control plane.

The concept of a "centralized control plane" is unclear. Perhaps this means "centralized controller"? Or you could be distinguishing between point-to-point protocols and a distributed control plane? But, of course, either control plane protocol approach could be used with distributed components or centralized components. And, I think, it is the components that are important for the architecture, not the protocol methods. This is confirmed by other paragraphs in the section.